// src/app/pages/admin/components/users/components/usuarios-table/user-response.interface.ts
export interface UserResponseInterface {
    idusuario?: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    idstatususuario: number;
    emailsms: string;
    statususuario?: string;
    rol?: string;
    descripcion?: string;
}