// src/app/pages/admin/components/users/components/usuarios-table/user.interface.ts
export interface UserInterface {
    claveauth?: string,
    nicknameauth?: string,
    usuarioauth?: string,
    idusuario?: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    idstatususuario: number;
    emailsms: string;
}